fix(docs): verify versioned builds and restore scoped search
