+++ .github/workflows/ci.yml
@@ -34,6 +34,9 @@ jobs:
       - name: Verify generated documentation contracts
         run: bun run docs:check
 
+      - name: Verify multiple documentation versions
+        run: bun --cwd tooling/docs check:versions
+
   typecheck:
     needs: build
     runs-on: ubuntu-latest
